// lib/interfaces.tsx
interface webcam {
  id: string;
  price: number | null;
  name: string | null;
  resolutions: string | null;
  connection: string | null;
  focus_type: string | null;
  fov: string | null;
  createdAt: string | null;
  updatedAt: string | null;
  manufacturer: string | null;
}